feat: cloudant da
