+++ solutions/fscloud/provider.tf
@@ -1,0 +1,5 @@
+provider "ibm" {
+  ibmcloud_api_key = var.ibmcloud_api_key
+  region           = var.region
+  ibmcloud_timeout = 60
+}
